+++ include/Dataset/DatasetStandard.h
@@ -6,8 +6,8 @@
 #include "Dataset.h"
 
 class DatasetStandard : public Dataset {
-    u32 inputs = 0; // input vector size
-    u32 outputs = 0; // output vector size
+    size_t inputs = 0; // input vector size
+    size_t outputs = 0; // output vector size
     size_t batchSize;
     size_t currentTrainSample = 0;
     size_t currentValSample = 0;
@@ -23,9 +23,9 @@ class DatasetStandard : public Dataset {
     std::pair<const Matrix2D &, const Matrix2D &> GetValidationSample(bool moveCursor) override;
     std::pair<const std::vector<Matrix2D> &, const std::vector<Matrix2D> &> GetValidationSamples() override;
     std::pair<const std::vector<Matrix2D> &, const std::vector<Matrix2D> &> GetTestSamples() override;
-    [[nodiscard]] u32 GetInputs() const override;
-    [[nodiscard]] u32 GetOutputs() const override;
-    [[nodiscard]] u32 GetBatchSize() const override;
+    [[nodiscard]] size_t GetInputs() const override;
+    [[nodiscard]] size_t GetOutputs() const override;
+    [[nodiscard]] size_t GetBatchSize() const override;
     void PreprocessMean() override;
 public:
     explicit DatasetStandard(const char* filename, size_t batchSize, f32 testCoef, f32 validationCoef);
